Доделал контроллеры, теперь надо понять какие методы нам нужны (регистрацию пока не стал, пока в ней нет необходимости)

+++ CRM.WebAPI/CRM.Infrastructure/Repositories/PaymentRepository.cs
@@ -1,4 +1,6 @@
 using System.Threading.Tasks;
+using CRM.Core.Implement;
+using CRM.Core.Interfaces;
 using CRM.Domain.Models;
 using CRM.Infrastructure.CreationObjectFromSQL;
 using CRM.Infrastructure.Interfaces;
@@ -13,18 +15,21 @@ public class PaymentRepository : RepositoryBase, IPaymentRepository
     {
     }
 
-    public async Task<IEnumerable<Payment>> GetAll()
+    public async Task<IOperationResult<IEnumerable<Payment>>> GetAll()
     {
-        return await GetDataSql<Payment, PaymentCreator>("SELECT * FROM payments");
+        return new Success<IEnumerable<Payment>>(await GetDataSql<Payment, PaymentCreator>("SELECT * FROM payments"));
     }
 
-    public async Task<Payment> GetById(Guid id)
+    public async Task<IOperationResult< Payment>> GetById(Guid id)
     {
-        return (await GetDataSql<Payment, PaymentCreator>("SELECT * FROM payments WHERE payment_id = @id",
-            new NpgsqlParameter("@id", id))).First();
+        var result = (await GetDataSql<Payment, PaymentCreator>("SELECT * FROM payments WHERE payment_id = @id",
+            new NpgsqlParameter("@id", id))).FirstOrDefault();
+        if (result == null)
+            return new ElementNotFound<Payment>(null,"Not found payment with current id!");
+        return new Success<Payment>(result);
     }
 
-    public async Task Put(Payment payment)
+    public async Task<IOperationResult<Guid>> Put(Payment payment)
     {
         var paymentId = Guid.NewGuid();
         await ExecuteSql(
@@ -35,10 +40,15 @@ public class PaymentRepository : RepositoryBase, IPaymentRepository
             new NpgsqlParameter("@paymentDate", payment.PaymentDate),
             new NpgsqlParameter("@paymentAmount", payment.PaymentAmount),
             new NpgsqlParameter("@paymentMethod", payment.PaymentMethod));
+        return new Success<Guid>(paymentId);
     }
 
-    public async Task Update(Payment dataToUpdate)
+    public async Task<IOperationResult> Update(Payment dataToUpdate)
     {
+        var paymentToUpdate = await GetById(dataToUpdate.PaymentId);
+        if (!paymentToUpdate.Successful)
+            return new ElementNotFound("Not found payment with current id");
+        
         await ExecuteSql(
             "UPDATE payments SET client_id = COALESCE(@clientId, client_id), visit_id = COALESCE(@visitId, visit_id), payment_date = COALESCE(@paymentDate, payment_date), payment_amount = COALESCE(@paymentAmount, payment_amount), payment_method = COALESCE(@paymentMethod, payment_method) WHERE payment_id = @id",
             new NpgsqlParameter("@clientId", dataToUpdate.ClientId),
@@ -47,10 +57,15 @@ public class PaymentRepository : RepositoryBase, IPaymentRepository
             new NpgsqlParameter("@paymentAmount", dataToUpdate.PaymentAmount),
             new NpgsqlParameter("@paymentMethod", dataToUpdate.PaymentMethod),
             new NpgsqlParameter("@id", dataToUpdate.PaymentId));
+        return new Success();
     }
 
-    public async Task RemoveById(Guid id)
+    public async Task<IOperationResult> RemoveById(Guid id)
     {
+        var paymentToDelete = await GetById(id);
+        if (!paymentToDelete.Successful)
+            return new ElementNotFound("Not found payment with current id");
         await ExecuteSql("DELETE FROM payments WHERE payment_id = @id", new NpgsqlParameter("@id", id));
+        return new Success();
     }
 }
